Consolidate variables into cache
