+++ x/pocketcore/types/cache.go
@@ -4,7 +4,9 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
+	"github.com/pokt-network/pocket-core/crypto"
 	"github.com/tendermint/tendermint/config"
+	"github.com/tendermint/tendermint/privval"
 	"log"
 	"path/filepath"
 	"sync"
@@ -31,7 +33,11 @@ var (
 
 	globalEvidenceSealedMap sync.Map
 
+	PkFromAddressMap map[string]crypto.PrivateKey
+
 	globalEvidenceSealedMapMap map[string]sync.Map
+
+	GlobalPVKeyFiles []privval.FilePVKey
 )
 
 // "CacheStorage" - Contains an LRU cache and a database instance w/ mutex
@@ -249,6 +255,33 @@ func GetSessionWithNodeAddress(header SessionHeader, address *sdk.Address) (sess
 		fmt.Println(fmt.Errorf("could not unmarshal into session from cache with header %v", header))
 	}
 	return
+}
+
+func InitializePkMap() error {
+	PkFromAddressMap = make(map[string]crypto.PrivateKey)
+	for _, pvKey := range GlobalPVKeyFiles {
+		pk, er := crypto.PrivKeyToPrivateKey(pvKey.PrivKey)
+		address := sdk.Address(pk.PublicKey().Address())
+		if er != nil {
+			return er
+		}
+		PkFromAddressMap[address.String()] = pk
+	}
+	return nil
+}
+
+func AddPVKeyFileMap(filePVKey privval.FilePVKey) {
+	GlobalPVKeyFiles = append(GlobalPVKeyFiles, filePVKey)
+}
+
+func GetPkFromAddress(address *sdk.Address) (crypto.PrivateKey, error) {
+	if PkFromAddressMap == nil {
+		err := InitializePkMap()
+		if err != nil {
+			return nil, err
+		}
+	}
+	return PkFromAddressMap[address.String()], nil
 }
 
 // "SetSession" - Sets a session (value) in the stores using the header (key)
