+++ x/pocketcore/types/config.go
@@ -3,7 +3,8 @@ package types
 import (
 	"encoding/hex"
 	"fmt"
-	"github.com/pokt-network/pocket-core/x/pocketcore/keeper"
+	"github.com/tendermint/tendermint/privval"
+
 	"sync"
 	"time"
 
@@ -38,12 +39,13 @@ func InitConfigOld(chains *HostedBlockchains, logger log.Logger, c types.Config)
 
 func InitConfig(chains *HostedBlockchains, logger log.Logger, c types.Config) {
 	cacheOnce.Do(func() {
+		GlobalPVKeyFiles = make([]privval.FilePVKey, 5)
 		globalEvidenceCache = new(CacheStorage)
 		globalSessionCache = new(CacheStorage)
 		globalEvidenceSealedMap = sync.Map{}
 		globalEvidenceCache.Init(c.PocketConfig.DataDir, c.PocketConfig.EvidenceDBName, c.TendermintConfig.LevelDBOptions, c.PocketConfig.MaxEvidenceCacheEntires, false)
 		globalSessionCache.Init(c.PocketConfig.DataDir, "", c.TendermintConfig.LevelDBOptions, c.PocketConfig.MaxSessionCacheEntries, true)
-		for key, _ := range keeper.PkFromAddressMap {
+		for key := range PkFromAddressMap {
 			globalEvidenceCacheMap[key] = new(CacheStorage)
 			globalEvidenceSealedMapMap[key] = sync.Map{}
 			globalSessionCacheMap[key] = new(CacheStorage)
@@ -95,7 +97,7 @@ func FlushSessionCache() {
 }
 
 func FlushSessionCacheAll() {
-	for k, _ := range keeper.PkFromAddressMap {
+	for k, _ := range PkFromAddressMap {
 		err := globalSessionCacheMap[k].FlushToDB()
 		if err != nil {
 			fmt.Printf("unable to flush sessions to the database before shutdown!! %s\n", err.Error())
